dom/console/tests/test_console_proto.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for console.__proto__</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script type="application/javascript">

  isnot(Object.getPrototypeOf(console), Object.prototype, "Foo");
  is(Object.getPrototypeOf(Object.getPrototypeOf(console)), Object.prototype, "Boo");

  </script>
</body>
</html>
